`ParserComparisonTest` inherits from `ParserComparisonTestInterface`
This allows us to create specialisms that implement custom
`getSnippets()` behaviour whilst sharing the core test plumbing.
